// src/data/actions/enum.ts
export enum action_types {
    USER_LOGGING_INIT = 'USER_LOGIN_INIT',
    USER_LOGGING_END = 'USER_LOGGING_END',
    USER_LOGIN = 'USER_LOGIN',
    USER_LOGIN_FAIL = 'USER_LOGIN_FAIL',
    USER_LOGOUT = 'USER_LOGOUT',
    RETRIEVE_REMINDERS = 'RETRIEVE_REMINDERS',
    UPDATE_REMINDERS = 'UPDATE_REMINDERS',
    FILTER_REMINDERS = 'FILTER_REMINDERS',
    CLEAR_FILTER_REMINDERS = 'CLEAR_FILTER_REMINDERS'
}